==> 3_Analise_exploratoria.Rmd <==
---
title: "Estatísticas descritivas: resumo das variáveis"
output:
    github_document:
     toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Tabela de estatísticas descritivas

```{r echo=TRUE, message=FALSE, warning=FALSE, fig.align='center', fig.height=3.5, fig.width=5}
## taxas

rm(list=ls())

library(broom)
library(sf)
library(tidyverse)
library(scales)

aed_tab <- read_rds("./Tabelas/base_parana.rds") %>%
  st_set_geometry(NULL) %>%
  mutate(prop_populacao_urbana = (prop_populacao_urbana)*100) %>%
  select(-seq(1:3)) %>%
  # compute broom::tidy summary of each var
  map(~tidy(summary(.x))) %>% 
  do.call(rbind, . ) %>%  # VRAU!
  # ajustando os números para um formáto publicável
  transmute(
     variavel = rownames(.),
    `Mínimo`  = minimum, 
     Q1       = q1,      
     Mediana  = median,  
     Q3       = q3,      
    `Máximo`  = maximum, 
    `Média`   = mean
    )

# vetor pivot
aed_tab$`Variável` <- aed_tab$variavel


# construindo um vetor com os nomes 'arrumados'
variavel <- aed_tab$variavel[-c(4:11, 24:57)]  %>%
  str_to_title() %>%
  # e.g.: "idhm_2010" -> "idhm" 
  str_replace_all("_2010", "") %>%
  str_replace_all("_", " ") %>%
  str_replace("aca", "aça") %>%
  str_replace("nca", "nça") %>%
  str_replace("çao", "ção") %>%
  str_replace("Mamografos ", "Mamógrafos ") %>%
  str_replace("\\ sus", "\\ SUS") %>%
  str_replace("Obitos", "Óbitos") %>%
  str_replace_all("População residente ", "População ") %>%
  str_replace_all("saude", "saúde") %>%
  str_replace_all("Indice de gini", "Índice de Gini") %>%
  str_replace_all("per capita", "(pc)") %>%
  str_replace_all("Idhm", "IDHM") %>%
  str_replace_all(" do ", ": ") %>%
  str_replace("Probabilidade de sobrevive", "Prob. de sobrevivê") %>%
  str_replace("ate ", "até ") %>%
  str_replace("Area km2", "Área $KM^2$")

# variaveis em sequência de faiza etária
obito <- c(paste("Óbitos (", seq(20, 75, by=10), "-", seq(29,79,by=10), ")", sep = ""),
                 "Óbitos (80+)",
                 "Óbitos tot. idade")

pop_fem <- c(paste("Pop. feminina (", seq(0,75,by=5), "-", seq(4,79,by=5), ")", sep=""), "Pop. feminina (80+)")
pop_msc <- c(paste("Pop. masculina (",seq(0,75,by=5), "-", seq(4,79,by=5), ")", sep=""), "Pop. masculina (80+)")

aed_tab$`Variável`[-c(4:11, 24:57)] <- variavel
aed_tab$`Variável`[4:11]  <- obito
aed_tab$`Variável`[24:57] <- c(pop_fem, pop_msc)

write_rds(aed_tab, "./Tabelas/aed_tab.rds")

aed_tab %>%
  select("Variável", c(2:7)) %>%
  knitr::kable(digits=3)

```


# Código para 'traduzir' e renderizar uma tabela do `R` para o LaTeX

## Um exemplo

```{r echo=TRUE, message=FALSE, warning=FALSE, fig.align='center', eval=FALSE}

library(kableExtra)

aed_tab %>%
  select("Variável", c(2:7)) %>%
  slice(c(1:57,74)) %>%
  mutate(`Média` = round(`Média`,1)) %>%
  kable(format = "latex",
       # align = 'c',
        format.args = list(decimal.mark=",", big.mark=".", scientific=FALSE),
        booktabs = T,
        longtable = T,
        caption = "Resumo das Variáveis - discretas",
        digit=0,
        escape = FALSE,
        linesep = "") %>%
  kable_styling(latex_options = c("striped", "repeat_header", "scale_down"))

aed_tab %>%
  select("Variável", c(2:7)) %>%
  slice(-c(1:57, 74)) %>%
  mutate(`Média` = round(`Média`,3)) %>%
    kable(format = "latex",
          digits = 3,
        #align = 'c',
        format.args = list(decimal.mark=",", big.mark=".", scientific=FALSE),
        booktabs = T,
        longtable = T,
        caption = "Resumo das Variáveis - contínuas",
        escape = FALSE,
        linesep = "") %>%
 kable_styling(latex_options = c("striped", "repeat_header", "scale_down"))
```

## População e Território no estado do Paraná

O código abaixo gera uma tabela em LaTeX e seu resultado é copiado para o arquivo `0_resultados.Rmd`:

```{r warning=FALSE, message=FALSE}



# é melhor copiar o output e colocar inline com a linguagem TEX
library(tidyverse)
library(readr)
library(kableExtra)
library(sf)

total <- read_rds("./Tabelas/base_parana.rds") %>%
  st_set_geometry(NULL) %>%
  group_by() %>%
  summarise(sum(populacao_total_2010),
            sum(populacao_rural_2010),
            sum(populacao_urbana_2010),
            sum(area_km2),
            sum(populacao_residente_masculina_2010),
            sum(populacao_residente_feminina_2010),
            sum(populacao_mulheres)) %>%
  mutate(dens = `sum(populacao_total_2010)`/`sum(area_km2)`,
         perc = round((`sum(populacao_urbana_2010)`/`sum(populacao_total_2010)`)*100)) %>%
  gather(value=Total) %>%
  mutate(index=c(1, 2, 3, 99, 6, 7, 8, 5, 4)) %>%
  arrange(index) %>%
  select(-key) %>%
  filter(index != 99)


read_rds("./Tabelas/aed_tab.rds") %>% 
  filter(str_detect(`Variável`, "opulação|Dens")) %>%
  
  mutate(index = c(1,2,3,6,7,5,4,8),
        `População Residente` = c("População Residente", #1
                                  "Pop. Rural",               #3
                                  "Pop. Urbana",              #4
                                  "Pop. Masculina",           #6
                                  "Pop. Feminina",            #7
                                  "Dens. Demográfica",        #2
                                  "Pop. Urbana (perc.)",      #5
                                  "Pop. Feminina ($15+$ anos)"#8
                                   )) %>%
  arrange(index) %>%
  select(9,2:7) %>%
  mutate(Total = total$Total) %>%
  kable(format = "latex",
        format.args = list(decimal.mark=",", big.mark=".", scientific=FALSE),
        booktabs  = T,
        digits    = 0,
        caption   = "Resumo - dados demográficos (2010)",
        escape    = FALSE) %>%
  kable_styling(latex_options = c("striped", "repeat_header", "scale_down", "hold_position")) %>%
  row_spec(1, bold = T) %>%
  group_rows("Território", 2,5) %>%
  group_rows("Gênero", 6,8)
  
```

## Neoplasia (câncer) maligna da mama, equipamentos e serviço de saúde.

O código abaixo gera uma tabela em LaTeX e seu resultado é copiado para o arquivo `0_resultados.Rmd`:

```{r warning=FALSE, message=FALSE, fig.align='center'}

library(kableExtra)

base_parana <- read_rds("./Tabelas/base_parana.rds") %>%
  filter(mamografos_existentes > 0) %>%
  mutate(mamografos_nao_sus = mamografos_existentes - mamografos_sus,
         municipio = as.character(municipio))

variav <- c( "Existentes",
             "Em uso",
             "Existentes e . pelo SUS",
             "Óbitos por câncer de mama\n(por 1.000 mulheres)",
             "N° de beneficiárias de planos de saúde privado\n(por 1.000 mulheres)")

zeros <- base_parana %>%
  st_set_geometry(NULL) %>%
  group_by() %>%
  summarise(sum(mamografos_existentes_taxa==0)/399,
            sum(mamografos_em_uso_taxa==0)/399,
            sum(mamografos_sus_taxa==0)/399,
            sum(obitos_total_idade_taxa==0)/399,
            sum(planos_saude_taxa==0)/399) %>%
  gather(value=zeros) %>%
  select(-key) %>%
  as_vector()


aed_tab %>%
  filter(str_detect(variavel, '_taxa')) %>%
  mutate(Indicador = variav,
         zeros     = paste(round(zeros*100,1), "\\%", sep="")) %>%
  select(9, 2:7, 10) %>%
  mutate_all(linebreak) %>%
  kable(format = "latex",
        format.args = list(decimal.mark=",", big.mark=".", scientific=FALSE),
        booktabs = T,
        digits   = 3,
        caption  = "Resumo - Indicadores de Saúde Pública",
        escape   = FALSE,
        linesep  = "") %>%
  kable_styling(latex_options = c("striped", "repeat_header", "scale_down", "hold_position"),
                font_size = 7) %>%
  group_rows("Equipamentos de mamografia\n(por 1.000 mulheres)", 1,3) %>%  
  group_rows("", 4,4) %>%
  group_rows("", 5,5)
 
```

## Equipamentos de Mamografia nos municípios

```{r eval=FALSE, echo=FALSE}
rm(list=ls())

library(kableExtra)
library(readr)
library(tidyverse)
library(sf)

diff(1:10, 2)

total <- function(x) {
  read_rds("./Tabelas/base_parana.rds") %>%
  st_set_geometry(NULL) %>%
  filter(x>0) %>%
  select(municipio, starts_with("mamogr"), -ends_with("_taxa")) %>%
  group_by()%>%
  summarise(mamografos_existentes=sum(mamografos_existentes),
            mamografos_sus=sum(mamografos_sus),
            mamografos_em_uso=sum(mamografos_em_uso)) %>%
  mutate(mamografos_existentes_taxa=NA,
         mamografos_sus_taxa=NA,
         mamografos_em_uso_taxa=NA,
         municipio= as.factor("TOTAL")) %>%
  select(municipio, everything())
}


checklist <- total('mamografos_em_uso') %>%
  bind_rows(., total('mamografos_sus')) %>%
  bind_rows(., total('mamografos_existentes')) %>%
  select(-municipio, -ends_with("_taxa")) %>%
  kable()


read_rds("./Tabelas/base_parana.rds") %>%
  st_set_geometry(NULL) %>%
  filter(mamografos_existentes>0) %>%
  select( #ordena coluna
    municipio,
    mamografos_existentes,
    mamografos_sus,
    mamografos_em_uso,
    mamografos_existentes_taxa,
    mamografos_sus_taxa,
    mamografos_em_uso_taxa
    ) %>%
  arrange(desc(mamografos_sus_taxa)) %>%
  bind_rows(total('mamografos_em_uso')) %>%
   kable(format = "latex",
        format.args = list(decimal.mark=",", big.mark=".", scientific=FALSE),
        booktabs  = T,
        digits    = 3,
        caption   = "Resumo - Equipamentos de mamografia",
        escape    = F,
        align     = c('l','c','c','c','c','c','c'),
        linesep = "") %>%
  kable_styling(latex_options = c("striped", "repeat_header", "scale_down", "hold_position"),
                font_size = 7) %>%
  add_header_above(c(" " = 1,  "Números Absolutos" = 3, "Taxas por 1.000 Mulheres" = 3), bold=T)
```

# Mapas

Dados:

```{r}

rm(list=ls())

library(broom)
library(sf)
library(tidyverse)
library(scales)

(base_parana <- read_rds("./Tabelas/base_parana.rds") %>%
    gather(situacao, taxa_de_mamografos,
          -c(cod_ibge:populacao_mulheres,
             obitos_total_idade_taxa:geometry),
          factor_key = T) %>%
   group_by(situacao) %>%
   nest())

```


```{r warning=FALSE, message=FALSE}

zeros <- base_parana$data %>%
  map(~ifelse(.x$taxa_de_mamografos>0, "não zero", "zero"))

base_parana$data <- map2(base_parana$data,
                         zeros,
                         ~mutate(.x, non_zero = .y))

com_mamogr <- base_parana$data %>%
  map(~st_set_geometry(.x, NULL)) %>%
  map(~filter(.x, non_zero=="não zero")) %>%
  map(~select(.x, municipio, mamografos_existentes, mamografos_sus, mamografos_em_uso))





#--
no_classes <- 6
quantiles <- com_mamogr %>%
  map(~{quantile(.$taxa_de_mamografos, probs = seq(0, 1, length.out = no_classes + 1))})
#--

titulos <- c("Existentes", "Em uso", "Disponíveis pelo SUS")

# gera os gráficos
mapas <- map2(base_parana$data, titulos, ~{
               ggplot(.x, aes(fill = ifelse(non_zero == "não zero", taxa_de_mamografos, NA))) +  # cria o ggplot
                geom_sf() +                    # desenha o mapa
                ggtitle(.y) +                  # adiciona o título
                theme(plot.title  = element_text(size = 12, hjust =.5),
                      axis.text.x = element_blank(),
                      axis.text.y = element_blank(),
                      axis.line.x = element_blank(),
                      axis.line.y = element_blank(),
                      panel.background = element_blank(),
                      panel.grid       = element_blank(),
                      panel.grid.major = element_line(colour = "white"),
                      axis.title.x = element_blank(),
                      axis.title.y = element_blank(),
                      axis.ticks   = element_blank(),
                      legend.position = "bottom") +
                # this is the main part
                scale_fill_viridis_c(option    = "magma",
                                     direction = -1,
                                     name      = "Taxa de mamógrafos por 1000 mulheres",
                                     # here we use guide_colourbar because it is still a continuous scale
                                     guide     = guide_colorbar(direction = "horizontal",
                                                                barheight = unit( 2, units = "mm"),
                                                                barwidth  = unit(50, units = "mm"),
                                                                draw.ulim = F,
                                                                title.position = 'top',
                                                                title.hjust = 0.5,
                                                                label.hjust = 0.5))
               
            })


```









Utilizamos o `ggpubr::ggarrange` para enquadrar as localidades.

```{r warning=FALSE, message=FALSE, fig.height=8, fig.width=10}
library(ggpubr)
mapas <- ggarrange(mapas[[1]], mapas[[2]], mapas[[3]],
                   ncol   = 2,
                   nrow   = 2,
                   align  = "hv",
                   legend = "bottom", 
                   common.legend = TRUE) %>%
  annotate_figure(.,
                 top = text_grob("Taxas de mamógrafos por 1000 Mulheres - Municípios do estado Paraná",
                                 color  = "black",
                                 vjust  = .5,
                                 size   = 14,
                                 family = "Times",
                                 just   = "center"))

mapas
```

# Outros gráficos


```{r warning=FALSE, message=FALSE, fig.align='center', echo=TRUE, fig.height=2.5, fig.width=4}
rm(list=ls())

library(readr)
library(tidyverse)
library(sf)
library(scales)

# HISTOGRAMA AREA Km^2
base_parana <- read_rds("./Tabelas/base_parana.rds")

base_parana %>%
ggplot(aes(x = log10(area_km2))) +
  geom_histogram(colour = "black", fill = NA, size = .2) +
  # as lihas média e mediana
  geom_vline(aes(xintercept = mean(log10(area_km2)),   color = "Média"  ), linetype = "dashed", size=.5) +
  geom_vline(aes(xintercept = median(log10(area_km2)), color = "Mediana"), linetype = "dashed", size=.5) +
  scale_color_manual(name = "", values = c("blue","red")) +
  # scale_x_continuous(labels = comma, breaks = seq(0, max(base_parana$area_km2), by=500000)) +
  # quadro-resumo
  geom_label(label=expression(atop("Área total: 199.307.988 " ~ Km^{2}, "n: 399")),
             x=3.2, y=33, parse=TRUE, size=2.2, label.size=.2) +
  # formatação básica
  labs(title = "Distribuição da área dos municípios do estado do Paraná",
       x = expression(log[10](Km^{2})),
       y = "Frequência") +
  theme(plot.title       = element_text(size=9, hjust = .35),
        plot.subtitle    = element_text(hjust = .5),
        legend.position  = c(.8,.6),
        legend.title     = element_blank(),
        legend.background= element_rect(line='solid', color='white', size=.2),
        legend.key.size  = unit(.2, "cm"),
        axis.ticks       = element_line(),
        legend.text      = element_text(size=7),
        axis.text.x      = element_text(size=7),
        axis.text.y      = element_text(size=7),
        axis.line.x      = element_line(),
        axis.line.y      = element_line(),
        panel.background = element_blank(),
        axis.title.x     = element_text(size=7),
        axis.title.y     = element_text(size=7))
```


```{r warning=FALSE, message=FALSE, fig.align='center', echo=TRUE, fig.height=4, fig.width=4}
rm(list=ls())

library(tidyverse)
require(sf)
library(readr)

mamografos <- read_rds("./Tabelas/base_parana.rds") %>%
  st_set_geometry(NULL) %>% 
  select(mamografos_existentes_taxa, mamografos_sus_taxa, mamografos_em_uso_taxa) %>%
  gather() %>%
  filter(value>0) %>%
  mutate(key = as.factor(key))

levels(mamografos$key) <- c("Em uso", "Existentes", "Disp. pelo SUS")



mamografos %>%
  ggplot(aes(y=log(value), x=1)) +
    geom_boxplot() +
    labs(title = "Distribuição dos Equipamentos de Mamografia nos municípios do estado do Paraná",
         subtitle="Logarítmo natural da taxa por 1.000 mulheres (20 anos e mais)",
         y = "Taxa",
         caption = "Exceto municípios sem equipamento de mamografia")+
    theme(plot.title = element_text(size=7),
          plot.subtitle = element_text(size=6),
          plot.caption = element_text(size=5),
          axis.text.x = element_blank(),
          axis.title.x = element_blank(),
          axis.title.y = element_text(size=6),
          axis.ticks.x = element_blank(),
          strip.text = element_text(size=5)) +
    facet_wrap(key ~ .)
```
