==> 4_Autocorrelacao_espacial.Rmd <==
---
title: "Autocorrelação Espacial"
output:
    github_document:
     toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



## Matriz de Contiguidade

A obtenção da matriz de vizinhança é feita à partir do objeto do tipo `sp`

```{r warning=FALSE, message=FALSE}
rm(list=ls())

library(tidyverse)
library(sf) 
library(readr)

sp_parana <- read_rds("./Tabelas/base_parana.rds") %>% as("Spatial")
class(sp_parana)
```

Foram definidas 8 tipos de matrizes diferentes:

* Vizinhança por fronteira: _queen_ e _rook_
* Vizinhança por distância: 5 a 10 vizinhos mais próximos

```{r warning=FALSE, message=FALSE}
library(spdep)
w_knear <- SpatialPoints(sp_parana)

# lista com matrizes de vizinhanças
lista_w <- list(w_queen = nb2listw(poly2nb(sp_parana, queen = T), style = "W"),
                w_rook  = nb2listw(poly2nb(sp_parana, queen = F), style = "W"),
                w_knear5  = nb2listw(knn2nb(knearneigh(w_knear, k = 5))),
                w_knear6  = nb2listw(knn2nb(knearneigh(w_knear, k = 6))),
                w_knear7  = nb2listw(knn2nb(knearneigh(w_knear, k = 7))),
                w_knear8  = nb2listw(knn2nb(knearneigh(w_knear, k = 8))),
                w_knear9  = nb2listw(knn2nb(knearneigh(w_knear, k = 9))),
                w_knear10 = nb2listw(knn2nb(knearneigh(w_knear, k = 10)))
                )
 
```

## [FIX] Microrregião de saúde:

```{r}
print("Inserir matriz de micorregioes de saúde")

library(spdep)
micro <- read.gal("./shp/w/micro.gal", override.id = TRUE)
summary(micro)
```



## Vizualizando contiguidade:

```{r echo=FALSE, warning=FALSE, message=FALSE, fig.width=20, fig.height=25}
coords <- coordinates(sp_parana)

par(mfrow = c(4, 2), mai =  c(.1, .1, .5, .1))
plot(sp_parana, main = "k = 5", border = "grey", lwd = 0.7)
plot(lista_w$w_knear5, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "k = 6", border = "grey", lwd = 0.7)
plot(lista_w$w_knear6, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "k = 7", border = "grey", lwd = 0.7)
plot(lista_w$w_knear7, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "k = 8", border = "grey", lwd = 0.7)
plot(lista_w$w_knear8, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "k = 9", border = "grey", lwd = 0.7)
plot(lista_w$w_knear9, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "k = 10", border = "grey", lwd = 0.7)
plot(lista_w$w_knear10, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "Matriz 'Qeen'", border = "grey", lwd = 0.7)
plot(lista_w$w_queen, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
plot(sp_parana, main = "Matriz 'Rook'", border = "grey", lwd = 0.5)
plot(lista_w$w_rook, coords, add = T, points = FALSE, col = "red", lwd = 0.7)
```


# Variáveis defasadas espacialmente

## Carregando as tabelas

Os mamógrafos são classificados por:

* Existentes: Soma de todos os mamógrafos existentes na localidade  
* Em uso: Subtrai os mamógrafos sem uso dos existentes
* Disponível pelo SUS: mamógrafos utilizados na rede pública de saúde

Para começar a análise, optamos pelo agrupamento dos dados nesas 3 categorias. A função `purrr::nest` faz esse trabalho aninhando uma base de dados para cada categoria.


```{r}


(base_parana <- read_rds("./Tabelas/base_parana.rds") %>% 
   select(-c(mamografos_existentes:mamografos_sus,
             obitos_20_a_29_anos:populacao_masculina_com_80_anos_e_mais_2010,
             67:71,
             idhm_renda_2010:area_km2)) %>%
   gather(situacao, key, -c(cod_ibge:municipio, geometry),
          factor_key = T) %>%
   group_by(situacao) %>%
   nest())
```

## Defasagem espacial (_lag_)

É construída através da função `spdep::lag.listw`


```{r}
# no cálculo da taxa, a função scale() padroniza os dados:
situacao <- map(base_parana$situacao, ~rep(.x, 399))


base_parana$data <- base_parana$data %>% 
   map(~mutate(.x, taxa = scale(key)))

# lista com taxas defasadass de acordo com a matriz de contiguidade
(base_parana <- map2(.x = base_parana$data,
                     .y = list(rep(lista_w, 3)),
                     .f = ~{mutate(.x, queen   = lag.listw(.y$w_queen, .x$taxa),
                                       rook    = lag.listw(.y$w_rook, .x$taxa),
                                       knear5  = lag.listw(.y$w_knear5, .x$taxa),
                                       knear6  = lag.listw(.y$w_knear6, .x$taxa),
                                       knear7  = lag.listw(.y$w_knear7, .x$taxa),
                                       knear8  = lag.listw(.y$w_knear8, .x$taxa),
                                       knear9  = lag.listw(.y$w_knear9, .x$taxa),
                                       knear10 = lag.listw(.y$w_knear10, .x$taxa),
                                   )}) %>%
                  map(.f = ~{st_set_geometry(., NULL)}) %>%
                  map(.f = ~{select(., cod_ibge, cod_munici, municipio, taxa:knear10)}) %>%
                  map2_df(.x = .,
                          .y = situacao,
                          .f = ~{mutate(.x, situacao = .y)}) %>%
                  group_by(situacao) %>%
                  nest() %>%
                  rename(lags = data) %>%
                  left_join(base_parana, ., by="situacao"))

```

## Cálculo do Índice de Moran



```{r}
library(broom)

(base_parana <- base_parana %>% 
    
      mutate(moran_queen   = map(base_parana$lags, ~lm(queen ~ taxa, data = .x)) %>% map(tidy), 
             moran_rook    = map(base_parana$lags, ~lm(rook ~ taxa, data = .x)) %>% map(tidy),
             moran_knear5  = map(base_parana$lags, ~lm(knear5 ~ taxa, data = .x))  %>% map(tidy),
             moran_knear6  = map(base_parana$lags, ~lm(knear6 ~ taxa, data = .x)) %>% map(tidy),
             moran_knear7  = map(base_parana$lags, ~lm(knear7 ~ taxa, data = .x)) %>% map(tidy),
             moran_knear8  = map(base_parana$lags, ~lm(knear8 ~ taxa, data = .x)) %>% map(tidy),
             moran_knear9  = map(base_parana$lags, ~lm(knear9 ~ taxa, data = .x)) %>% map(tidy),
             moran_knear10 = map(base_parana$lags, ~lm(knear10 ~ taxa, data = .x)) %>% map(tidy)
             
              ))
```


## Diagrama de dispersão de Moran

### Matriz _queen_

```{r  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}

plot_my_data <- function(mydata, y) {
    ggplot(mydata) + 
      geom_point(aes_string(x = 'taxa', y = y), 
                 shape  = 21,
                 fill   = "white",
                 size   = 1.2,
                 stroke =  .6) + 
      
      theme_bw(base_size = 8) +
      theme(plot.title       = element_text(hjust = .5),
            plot.subtitle    = element_text(hjust = .5, margin = margin(b = -10)),
            axis.text        = element_text(colour = "black"),
            axis.text.x      = element_text(size = 6.5),
            axis.text.y      = element_text(size = 6.5),
            axis.ticks       = element_line(size =  .3),
            axis.line        = element_blank(),
            axis.title.x     = element_text(size = 10),
            axis.title.y     = element_text(size = 10),
            panel.background = element_rect(size = .3),
            panel.grid = element_blank()) +
              
      labs(x = "lag Variável",
           y = "Variável") +
  
      scale_y_continuous(limits = c(-.6, 2.6),
                         breaks = seq(-.5, 2.5, by = 0.5)) +
      scale_x_continuous(limits = c(-1.1, 11),
                         breaks = seq(-1, 11, by = 1)) +  
  
      geom_vline(xintercept = 0, size = .3) +
      geom_hline(yintercept = 0, size = .3) }

titulos <- base_parana$situacao
 
library(ggpubr)
queen <- base_parana$lags %>% 
   map2(.x = .,c(rep("queen", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_queen, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                       intercept = .y$estimate[[2]],
                                                       size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})




ggarrange(queen[[1]],  queen[[2]],  queen[[3]],  queen[[4]],
          queen[[1]],  queen[[2]],  queen[[7]],  queen[[8]],
          queen[[9]],  queen[[10]], queen[[11]], queen[[12]],
          queen[[12]], queen[[14]], queen[[15]], queen[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - Matriz de Contiguidade Queen",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### Matriz _rook_

```{r  echo=TRUE, warning=FALSE, message=FALSE, fig.height=15, fig.width=10}



rook <- base_parana$lags %>% 
   map2(.x = ., c(rep("rook", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_rook, ~{.x + geom_abline(slope     = .y$estimate[[1]],
                                                      intercept = .y$estimate[[2]],
                                                      size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(rook[[1]],  rook[[2]],  rook[[3]],  rook[[4]],
          rook[[5]],  rook[[6]],  rook[[7]],  rook[[8]],
          rook[[9]],  rook[[10]], rook[[11]], rook[[12]],
          rook[[13]], rook[[14]], rook[[15]], rook[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - Matriz de Contiguidade Rook",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### 5 Vizinhos mais próximos

```{r echo=FALSE, warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear5 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear5", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear5, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                        intercept = .y$estimate[[2]],
                                                        size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(knear5[[1]],  knear5[[2]],  knear5[[3]],  knear5[[4]],
          knear5[[5]],  knear5[[6]],  knear5[[7]],  knear5[[8]],
          knear5[[9]],  knear5[[10]], knear5[[11]], knear5[[12]],
          knear5[[13]], knear5[[14]], knear5[[15]], knear5[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```


### 6 vizinhos mais próximos

```{r  echo=FALSE,  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear6 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear6", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear6, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                        intercept = .y$estimate[[2]],
                                                        size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(knear6[[1]],  knear6[[2]],  knear6[[3]],  knear6[[4]],
          knear6[[5]],  knear6[[6]],  knear6[[7]],  knear6[[8]],
          knear6[[9]],  knear6[[10]], knear6[[11]], knear6[[12]],
          knear6[[13]], knear6[[14]], knear6[[15]], knear6[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### 7 vizinhos mais próximos

```{r  echo=FALSE,  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear7 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear7", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear7, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                        intercept = .y$estimate[[2]],
                                                        size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})
ggarrange(knear7[[1]],  knear7[[2]],  knear7[[3]],  knear7[[4]],
          knear7[[5]],  knear7[[6]],  knear7[[7]],  knear7[[8]],
          knear7[[9]],  knear7[[10]], knear7[[11]], knear7[[12]],
          knear7[[13]], knear7[[14]], knear7[[15]], knear7[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### 8 vizinhos mais próximos

```{r  echo=FALSE,  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear8 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear8", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear8, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                        intercept = .y$estimate[[2]],
                                                        size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(knear8[[1]],  knear8[[2]],  knear8[[3]],  knear8[[4]],
          knear8[[5]],  knear8[[6]],  knear8[[7]],  knear8[[8]],
          knear8[[9]],  knear8[[10]], knear8[[11]], knear8[[12]],
          knear8[[13]], knear8[[14]], knear8[[15]], knear8[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### 9 vizinhos mais próximos

```{r  echo=FALSE,  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear9 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear9", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear9, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                        intercept = .y$estimate[[2]],
                                                        size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(knear9[[1]],  knear9[[2]],  knear9[[3]],  knear9[[4]],
          knear9[[5]],  knear9[[6]],  knear9[[7]],  knear9[[8]],
          knear9[[9]],  knear9[[10]], knear9[[11]], knear9[[12]],
          knear9[[13]], knear9[[14]], knear9[[15]], knear9[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```

### 10 vizinhos mais próximos

```{r  echo=FALSE,  warning=FALSE, message=FALSE, fig.height=15, fig.width=10}


knear10 <- base_parana$lags %>% 
   map2(.x = .,c(rep("knear10", 16)), ~{plot_my_data(.x, .y)}) %>% 
   map2(., titulos, ~{.x + ggtitle(.y)}) %>%
   map2(., base_parana$moran_knear10, ~{.x + geom_abline(slope = .y$estimate[[1]],
                                                         intercept = .y$estimate[[2]],
                                                         size = .5, linetype = 'dashed')}) #%>%
#   map(. , ~{ggarrange(.x, ncol = 3, nrow   = 1, align  = "hv")})

ggarrange(knear10[[1]],  knear10[[2]],  knear10[[3]],  knear10[[4]],
          knear10[[5]],  knear10[[6]],  knear10[[7]],  knear10[[8]],
          knear10[[9]],  knear10[[10]], knear10[[11]], knear10[[12]],
          knear10[[13]], knear10[[14]], knear10[[15]], knear10[[16]],
          ncol   = 4,
          nrow   = 4,
          align  = "hv",
          legend = "bottom",
          common.legend = TRUE) %>%
  annotate_figure(.,
                  top = text_grob("Indicador de Moran - 5 vizinhos mais próximos",
                                  color  = "black",
                                  vjust  = .5,
                                  size   = 14,
                                  family = "Times",
                                  just   = "center"))
```
